Sends slack event on court date notification
