+++ apps/judicial-system/api/src/app/app.module.ts
@@ -15,7 +15,6 @@ import {
   InstitutionModule,
   CourtModule,
   FeatureModule,
-  EventModule,
 } from './modules/'
 
 const debug = !environment.production
@@ -47,7 +46,6 @@ const autoSchemaFile = environment.production
     CourtModule,
     FeatureModule,
     CmsTranslationsModule,
-    EventModule,
   ],
   providers: [BackendAPI],
 })
