+++ apps/judicial-system/api/src/app/modules/case/case.module.ts
@@ -1,10 +1,8 @@
 import { Module } from '@nestjs/common'
-import { EventModule } from '../event'
 
 import { CaseResolver } from './case.resolver'
 
 @Module({
-  imports: [EventModule],
   providers: [CaseResolver],
 })
 export class CaseModule {}
