+++ apps/judicial-system/api/src/app/modules/index.ts
@@ -5,4 +5,3 @@ export { CaseModule } from './case'
 export { FileModule } from './file'
 export { CourtModule } from './court'
 export { FeatureModule } from './feature'
-export { EventModule } from './event'
