+++ apps/judicial-system/api/src/environments/environment.ts
@@ -32,9 +32,6 @@ const devConfig = {
   features: {
     hidden: '',
   },
-  events: {
-    url: process.env.EVENT_URL,
-  },
 }
 
 if (process.env.NODE_ENV === 'production') {
@@ -119,9 +116,6 @@ const prodConfig = {
   },
   features: {
     hidden: process.env.HIDDEN_FEATURES,
-  },
-  events: {
-    url: process.env.EVENT_URL,
   },
 }
