+++ apps/judicial-system/backend/infra/judicial-system-backend.ts
@@ -66,6 +66,7 @@ export const serviceSetup = (): ServiceBuilder<'judicial-system-backend'> =>
       XROAD_CLIENT_PEM: '/k8s/judicial-system/XROAD_CLIENT_PEM',
       SECRET_TOKEN: '/k8s/judicial-system/SECRET_TOKEN',
       CONTENTFUL_ACCESS_TOKEN: '/k8s/judicial-system/CONTENTFUL_ACCESS_TOKEN',
+      EVENT_URL: '/k8s/judicial-system/EVENT_URL',
     })
     .initContainer({
       containers: [{ command: 'npx', args: ['sequelize-cli', 'db:migrate'] }],
