+++ apps/judicial-system/backend/src/app/modules/event/index.ts
