+++ apps/judicial-system/backend/src/app/modules/index.ts
@@ -4,3 +4,4 @@ export { CaseModule } from './case'
 export { NotificationModule } from './notification'
 export { FileModule } from './file'
 export { CourtModule } from './court'
+export { EventModule } from './event'
