+++ charts/judicial-system/values.staging.yaml
@@ -59,7 +59,6 @@ judicial-system-api:
     CONFIGCAT_SDK_KEY: /k8s/configcat/CONFIGCAT_SDK_KEY
     CONTENTFUL_ACCESS_TOKEN: /k8s/judicial-system/CONTENTFUL_ACCESS_TOKEN
     COURTS_CREDENTIALS: /k8s/judicial-system/COURTS_CREDENTIALS
-    EVENT_URL: /k8s/judicial-system/EVENT_URL
     SECRET_TOKEN: /k8s/judicial-system/SECRET_TOKEN
     XROAD_CLIENT_CERT: /k8s/judicial-system/XROAD_CLIENT_CERT
     XROAD_CLIENT_KEY: /k8s/judicial-system/XROAD_CLIENT_KEY
@@ -136,6 +135,7 @@ judicial-system-backend:
     EMAIL_FROM_NAME: /k8s/judicial-system/EMAIL_FROM_NAME
     EMAIL_REPLY_TO: /k8s/judicial-system/EMAIL_REPLY_TO
     EMAIL_REPLY_TO_NAME: /k8s/judicial-system/EMAIL_REPLY_TO_NAME
+    EVENT_URL: /k8s/judicial-system/EVENT_URL
     NOVA_PASSWORD: /k8s/judicial-system/NOVA_PASSWORD
     NOVA_URL: /k8s/judicial-system/NOVA_URL
     NOVA_USERNAME: /k8s/judicial-system/NOVA_USERNAME
